test: Order Service 단위테스트 작성, 엔티티 형변환(재고, 수량관련 Long -> Integer) - UserServiceUnitTest, ProductServiceUnitTest, OrderServiceUnitTest

+++ src/main/java/kr/hhplus/be/server/product/application/dto/ProductResponse.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.product.application.dto;
 
 import kr.hhplus.be.server.product.domain.entity.Product;
 
-public record ProductResponse(Long id, String name, Long price, Long stock) {
+public record ProductResponse(Long id, String name, Integer price, Integer stock) {
     public static ProductResponse from(Product product) {
         return new ProductResponse(product.getId(), product.getName(), product.getPrice(), product.getStock());
     }
